moves reporting system props to config package
