+++ migrate-common/src/main/java/org/opentestsystem/rdw/migrate/common/config/ReportingSystemProperties.java
@@ -1,4 +1,4 @@
-package org.opentestsystem.rdw.migrate.common.configuration;
+package org.opentestsystem.rdw.migrate.common.config;
 
 /**
  * Reporting configuration
