+++ migrate-common/src/main/java/org/opentestsystem/rdw/migrate/common/config/ReportingSystemPropertiesImpl.java
@@ -1,4 +1,4 @@
-package org.opentestsystem.rdw.migrate.common.configuration;
+package org.opentestsystem.rdw.migrate.common.config;
 
 import com.google.common.base.Objects;
