+++ migrate-common/src/main/java/org/opentestsystem/rdw/migrate/common/config/ReportingSystemPropertiesResolver.java
@@ -1,4 +1,4 @@
-package org.opentestsystem.rdw.migrate.common.configuration;
+package org.opentestsystem.rdw.migrate.common.config;
 
 import org.opentestsystem.rdw.multitenant.TenantKeyResolver;
 import org.slf4j.Logger;
