+++ migrate-olap/src/main/java/org/opentestsystem/rdw/ingest/migrate/olap/TenantConfiguration.java
@@ -1,7 +1,7 @@
 package org.opentestsystem.rdw.ingest.migrate.olap;
 
-import org.opentestsystem.rdw.migrate.common.configuration.ReportingSystemPropertiesResolver;
-import org.opentestsystem.rdw.migrate.common.configuration.ReportingSystemSettings;
+import org.opentestsystem.rdw.migrate.common.config.ReportingSystemPropertiesResolver;
+import org.opentestsystem.rdw.migrate.common.config.ReportingSystemSettings;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.cloud.context.config.annotation.RefreshScope;
 import org.springframework.context.annotation.Bean;
