+++ migrate-olap/src/main/java/org/opentestsystem/rdw/ingest/migrate/olap/repository/impl/JdbcWarehouseSqlListExecutionRepository.java
@@ -1,6 +1,6 @@
 package org.opentestsystem.rdw.ingest.migrate.olap.repository.impl;
 
-import org.opentestsystem.rdw.migrate.common.configuration.ReportingSystemProperties;
+import org.opentestsystem.rdw.migrate.common.config.ReportingSystemProperties;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.jdbc.core.PreparedStatementCallback;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
