+++ migrate-reporting/src/main/java/org/opentestsystem/rdw/ingest/migrate/reporting/TenantConfiguration.java
@@ -1,7 +1,7 @@
 package org.opentestsystem.rdw.ingest.migrate.reporting;
 
-import org.opentestsystem.rdw.migrate.common.configuration.ReportingSystemPropertiesResolver;
-import org.opentestsystem.rdw.migrate.common.configuration.ReportingSystemSettings;
+import org.opentestsystem.rdw.migrate.common.config.ReportingSystemPropertiesResolver;
+import org.opentestsystem.rdw.migrate.common.config.ReportingSystemSettings;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.cloud.context.config.annotation.RefreshScope;
 import org.springframework.context.annotation.Bean;
