+++ migrate-reporting/src/main/java/org/opentestsystem/rdw/ingest/migrate/reporting/repository/impl/JdbcSqlCopyWarehouseToStagingRepository.java
@@ -1,6 +1,6 @@
 package org.opentestsystem.rdw.ingest.migrate.reporting.repository.impl;
 
-import org.opentestsystem.rdw.migrate.common.configuration.ReportingSystemProperties;
+import org.opentestsystem.rdw.migrate.common.config.ReportingSystemProperties;
 import org.opentestsystem.rdw.ingest.migrate.reporting.repository.SqlCopyWarehouseToStagingRepository;
 import org.opentestsystem.rdw.ingest.migrate.reporting.step.SqlCopyStatements;
 import org.springframework.beans.factory.annotation.Autowired;
